sticky platforms work
lolol
