+++ TimePunch/Assets/Scripts/StickyPlatform.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 900c6c0f8b6d17a4b8bb7ffc1c30e9c1
+timeCreated: 1523846271
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
